Divided spetrafit into submodules
